requirements.txt
aiohttp
slackclient>=2.0.0
certifi
flask-wtf
flask
python-dotenv
flask-bootstrap
requests
